requirements.txt
numpy
omegaconf
msgspec
rouge-score
sacrebleu
sentencepiece
tiktoken
fsspec
blobfile
wandb
viztracer
lm-eval
scipy
pynvml
datatrove
orjson
transformers>=4.36.0
Pillow
matplotlib